refactor(command): move FsWatcher into feature/fs_watcher.rs
